rename: 폴더명 수정 savingsProduct -> savings-product
